% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_eDNA_tracks_functions.R
\name{check_time}
\alias{check_time}
\title{check_time}
\usage{
check_time(data)
}
\arguments{
\item{data}{A dataframe obtained from the function "spygen_waypoint". It must contain the following columns : "spygen_code", "date", "latitude_start_DD", "longitude_start_DD", "latitude_end_DD", "longitude_end_DD", "project", "time_start", "Time_waypoint_start", "Time_waypoint_end".}
}
\value{
A dataframe with a column named "diff_time_waypoints", indicating the difference in time (minutes) between the waypoint end and start.
}
\description{
This function calculate the difference in time (minutes) between the waypoint end and start.
}
