% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_eDNA_tracks_functions.R
\name{load_tracks}
\alias{load_tracks}
\title{load_tracks}
\usage{
load_tracks(path)
}
\arguments{
\item{path}{a character indicating the path to the waypoints.}
}
\value{
a dataframe of tracks coordinates.
}
\description{
This function load tracks from gps data.
}
