% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_eDNA_tracks_functions.R
\name{load_waypoint}
\alias{load_waypoint}
\title{load_waypoint}
\usage{
load_waypoint(path)
}
\arguments{
\item{path}{a character indicating the path to the waypoints.}
}
\value{
a dataframe of waypoints coordinates.
}
\description{
This function load waypoints from gps data.
}
