% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_eDNA_tracks_functions.R
\name{shapefile_tracks}
\alias{shapefile_tracks}
\title{shapefile_tracks}
\usage{
shapefile_tracks(eDNA_tracks, path_save)
}
\arguments{
\item{eDNA_tracks}{A dataframe, the "tracks_good_distance" one obtained with the "spygen_tracks" function.}

\item{path_save}{A character indicating the path to save data.}
}
\value{
Save the shapefile to the path indicated by \code{path_save}
}
\description{
This function convert tracks point as a dataframe into a shapefile.
}
