% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_clean_eDNA_functions.R
\name{species_clean_function}
\alias{species_clean_function}
\title{species_clean_function}
\usage{
species_clean_function(spygen_matrix, path_save)
}
\arguments{
\item{spygen_matrix}{a dataframe of species eDNA in the format site X species}
}
\value{
a list containing three objects :

A dataframe in the format site X species with new species names checked from FishBase;
A dataframe in the format site X species with old species names before checking from FishBase;
A character vector listing all removed species.
}
\description{
This function clean species eDNA data in terms of species names.
}
\details{
It removes all misidentified species (missing names, identified at the family level or as spp., sp., all species not spelled in the binomial format).

Once misidentified species has been removed, the function check the correct names of remaining species from FishBase.

Caution : this function does not check for species distribution..................
}
