% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_clean_eDNA_functions.R
\name{spygen_subset_function}
\alias{spygen_subset_function}
\title{spygen_subset_function}
\usage{
spygen_subset_function(eDNA_species_data_path, spygen_code_subset)
}
\arguments{
\item{eDNA_species_data_path}{a character indicating the path of cleaned spygen data. Data has to be in the format ".csv".}

\item{spygen_code_subset}{a character or a dataframe indicating which spygen_code to subset the data with. If supply a dataframe, make sure to name the spygen code column "spygen_code"}
}
\value{
a dataframe in the format site X species
}
\description{
This function create a subset of eDNA data by spygen_code.
}
