% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_eDNA_tracks_functions.R
\name{spygen_tracks}
\alias{spygen_tracks}
\title{spygen_tracks}
\usage{
spygen_tracks(waypoints, gps_tracks, distance_threshold, path_save)
}
\arguments{
\item{waypoints}{a dataframe, the "good_distance" one obtained with the "spygen_waypoint" function.}

\item{gps_tracks}{a dataframe of tracks coordinates.}

\item{distance_threshold}{a numeric indicating the maximal distance threshold in meter between a spygen survey coordinates and the closest tracks coordinates.}

\item{path_save}{a character indicating the path to save data.}
}
\value{
a list of three dataframe :
"high distance" contain spygen survey with distance to the closest tracks > \code{distance_threshold};
"good distance" contain spygen survey with distance to the closest tracks < \code{distance_threshold};
"na survey" contain spygen survey with no attributed tracks.
}
\description{
This function attributes to each spygen survey it's closest tracks at the date of the survey.
}
