% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_eDNA_tracks_functions.R
\name{spygen_waypoint}
\alias{spygen_waypoint}
\title{spygen_waypoint}
\usage{
spygen_waypoint(eDNA_metadata_path, waypoints, distance_threshold, path_save)
}
\arguments{
\item{eDNA_metadata_path}{a character indicating eDNA metadata path. These metadata includes spygen survey coordinates and dates.}

\item{waypoints}{a dataframe of waypoints coordinates.}

\item{distance_threshold}{a numeric indicating the maximal distance threshold in meter between a spygen survey coordinates and the closest waypoint coordinates.}

\item{path_save}{a character indicating the path to save data.}
}
\value{
a list of three dataframe :
"high distance" contain spygen survey with distance to the closest waypoints > \code{distance_threshold};
"good distance" contain spygen survey with distance to the closest waypoints < \code{distance_threshold};
"na survey" contain spygen survey with no attributed waypoints.
}
\description{
This function attributes to each spygen survey it's closest waypoint at the date of the survey.
}
