% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_clean_eDNA_functions.R
\name{convert_to_matrix_function}
\alias{convert_to_matrix_function}
\title{convert_to_matrix_function}
\usage{
convert_to_matrix_function(raw_spygen_path)
}
\arguments{
\item{raw_spygen_path}{A character indicating the path of spygen raw data. Data has to be in the format ".xlsx".}
}
\value{
A dataframe in the format site X species
}
\description{
This function convert spygen raw data into a site X species matrix.
}
