% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_clean_eDNA_functions.R
\name{species_clean_function}
\alias{species_clean_function}
\title{species_clean_function}
\usage{
species_clean_function(spygen_matrix, keep = NULL)
}
\arguments{
\item{spygen_matrix}{A dataframe of species eDNA in the format site X species.}

\item{keep}{A character of the type of species to keep while not being in the binominal format. It can be of the following: "Missing", "Empty", "Complex", "Family",
"Genus or higher", "Subspecies", "spp.", "sp.", "(cf)", "hybrid", "Unidentified", "?", "!", "/", "New." You can specify multiple types of species to keep.}
}
\value{
A list containing three objects :

A dataframe in the format site X species with new species names checked from FishBase;

A dataframe in the format site X species with old species names before checking from FishBase;

A character vector listing all removed species.
}
\description{
This function clean species eDNA data in terms of species names.
}
\details{
The function removes all misidentified species (missing names, identified at the family level or as spp., sp., all species not spelled in the binomial format).

Once misidentified species has been removed, the function check the correct names of remaining species from FishBase.

Caution : this function only convert data to a suitable format for analysis, with only basic cleaning step. This does not exempt users from checking the list of species returned by the functions (e.g., species detected outside their distribution range).
}
